Make an other bundle with the WpsServer implementation for OrbisGIS.
