+++ bundles/wps/wps-client/src/main/java/org/orbisgis/wpsclient/impl/WpsClientImpl.java
@@ -53,13 +53,11 @@ import org.orbisgis.sif.docking.DockingManager;
 import org.orbisgis.sif.docking.DockingPanel;
 import org.orbisgis.sif.docking.DockingPanelParameters;
 import org.orbisgis.sif.edition.EditorDockable;
-import org.orbisgis.sif.edition.EditorFactory;
 import org.orbisgis.sif.edition.EditorManager;
 import org.orbisgis.wpsclient.api.InternalWpsClient;
 import org.orbisgis.wpsclient.api.WpsClient;
 import org.orbisgis.wpsclient.api.utils.ProcessExecutionType;
 import org.orbisgis.wpsclient.impl.dataui.DataUIManager;
-import org.orbisgis.wpsclient.impl.editor.process.ProcessEditorFactory;
 import org.orbisgis.wpsclient.impl.utils.ToolBoxIcon;
 import org.orbisgis.wpsclient.api.utils.WpsJobStateListener;
 import org.orbisgis.wpsclient.impl.editor.log.LogEditableElement;
@@ -68,7 +66,7 @@ import org.orbisgis.wpsclient.impl.utils.Job;
 import org.orbisgis.wpsclient.impl.editor.process.ProcessEditableElement;
 import org.orbisgis.wpsclient.impl.editor.process.ProcessEditor;
 import org.orbisgis.wpsservice.model.*;
-import org.orbisgis.wpsservice.LocalWpsServer;
+import org.orbisgis.wpsserviceorbisgis.LocalWpsServer;
 import org.orbisgis.wpsservice.utils.ProcessMetadata;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
