+++ bundles/wps/wps-client/src/main/java/org/orbisgis/wpsclient/impl/dataui/JDBCTableFieldUI.java
@@ -46,7 +46,7 @@ import org.orbisgis.sif.components.renderers.JPanelListRenderer;
 import org.orbisgis.wpsclient.impl.WpsClientImpl;
 import org.orbisgis.wpsclient.api.dataui.DataUI;
 import org.orbisgis.wpsclient.impl.utils.ToolBoxIcon;
-import org.orbisgis.wpsservice.LocalWpsServer;
+import org.orbisgis.wpsserviceorbisgis.LocalWpsServer;
 import org.orbisgis.wpsservice.model.*;
 import org.xnap.commons.i18n.I18n;
 import org.xnap.commons.i18n.I18nFactory;
