+++ bundles/wps/wps-service-scripts/src/main/java/org/orbisgis/wpsservicescripts/WpsScriptsPackage.java
@@ -39,8 +39,8 @@ package org.orbisgis.wpsservicescripts;
 import net.opengis.ows._2.CodeType;
 import org.apache.commons.io.IOUtils;
 import org.orbisgis.frameworkapi.CoreWorkspace;
-import org.orbisgis.wpsservice.LocalWpsServer;
 
+import org.orbisgis.wpsservice.WpsServer;
 import org.orbisgis.wpsservice.controller.process.ProcessIdentifier;
 import org.orbisgis.wpsclient.api.InternalWpsClient;
 import org.slf4j.Logger;
@@ -102,7 +102,7 @@ public class WpsScriptsPackage {
      * The WPS service of OrbisGIS.
      * The WPS service contains all the declared processes available for the client (in OrbisGIS the toolbox).
      */
-    protected LocalWpsServer localWpsService;
+    protected WpsServer wpsServer;
 
     /**
      * The WPS client of OrbisGIS.
@@ -153,7 +153,7 @@ public class WpsScriptsPackage {
             LOGGER.error(I18N.tr("Unable to copy the content of the script to the temporary file."));
             return;
         }
-        List<ProcessIdentifier> piList = localWpsService.addLocalSource(tempFile,
+        List<ProcessIdentifier> piList = wpsServer.addLocalSource(tempFile,
                 icons,
                 false,
                 path);
@@ -172,7 +172,7 @@ public class WpsScriptsPackage {
      */
     protected void removeAllScripts(){
         for(CodeType idProcess : listIdProcess){
-            localWpsService.removeProcess(URI.create(idProcess.getValue()));
+            wpsServer.removeProcess(URI.create(idProcess.getValue()));
         }
     }
