+++ bundles/wps/wps-service/src/test/java/org/orbisgis/wpsservice/WpsServerImplTest.java
@@ -72,7 +72,7 @@ public class WpsServerImplTest {
      */
     @Before
     public void initialize(){
-        LocalWpsServerImpl localWpsServer = new LocalWpsServerImpl();
+        WpsServerImpl localWpsServer = new WpsServerImpl();
         localWpsServer.init();
 
         try {
